Refactored. - Simplified project structure and namespaces.
