+++ src/Client/Service.Model/DeleteInstance.cs
@@ -1,13 +1,9 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System;
+using Newtonsoft.Json;
 
 namespace OnlineManagementApiClient.Service.Model
 {
-    public class DeleteInstanceRequest
+    public class DeleteInstance
     {
         public Guid InstanceId { get; set; }
