+++ src/Client/Service.Model/GetOperationStatus.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace OnlineManagementApiClient.Service.Model
+{
+    public class GetOperationStatus
+    {
+        public Guid OperationId { get; set; }
+    }
+}
