--- src/Client/Service.Model/GetOperationStatusRequest.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OnlineManagementApiClient.Service.Model
-{
-    public class GetOperationStatusRequest
-    {
-        public Guid OperationId { get; set; }
-    }
-}
