+++ src/Client/Service.Model/OperationStatus.ItemDescription.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace OnlineManagementApiClient.Service.Model
+﻿namespace OnlineManagementApiClient.Service.Model
 {
     public class ItemDescription
     {
