+++ src/Client/Utility/ConsoleLogService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OnlineManagementApiClient.Service
+namespace OnlineManagementApiClient.Utility
 {
     public class ConsoleLogService : ILog
     {
