+++ src/Client/Utility/FileLogService.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OnlineManagementApiClient.Service
+namespace OnlineManagementApiClient.Utility
 {
     public class FileLogService : ILog
     {
